Patch adb code to work with Android 15 QPR1 with legacy Function FS
Starting with Android 14 QPR1, the newer adb doesn't work with our
Function FS patches.  Thus the Android 14 QPR1 branch was made so that older
devices had working adb. Thanks to acroreiser.

With Android 15, the following additional patches are needed so that
Function FS continues on older devices to get adb working (that is, adb
logcat).